% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_master.R
\name{lib_master}
\alias{lib_master}
\title{Installation and Loading of Required Libraries}
\usage{
lib_master(req_file, loc_lib = NULL)
}
\arguments{
\item{req_file}{Path to file with list of libraries. Plain-text, one
library per line.}

\item{loc_lib}{(Optional) Determine the folder, where the local library
should be installed. Defaults to the first element of .libPaths().}
}
\description{
Based on a plain-text file, this function installs and loads
all libraries and packages.
}
\examples{
req_file <- "requirements.txt"
writeLines("dplyr", req_file)

lib_master(req_file)
}
\author{
Sebastian Schweer
}
\keyword{init}
\keyword{libraries,}
