% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_init.R
\name{script_init}
\alias{script_init}
\title{Initialization of the Logger for Scripts}
\usage{
script_init(code_dir = ".", log_path = "logfile.txt", logger = "main",
  script_name = NULL)
}
\arguments{
\item{code_dir}{Directory in which the script is located. Defaults to "."}

\item{log_path}{Path of the logfile to be created. Defaults to "logfile.txt"
within the working directory.}

\item{logger}{Name of the logger handler to be initialized, defaults to "main".}

\item{script_name}{(Optional) Name of the Script or project at the head of
the log.}
}
\description{
This function initialzies the logging of the current script in a text
file. The head of the logfile contains essential information on the
build of the system, the R version, and the versions of the attached
packages.
}
\examples{
script_init()
logging::loginfo("This is not in logfile.txt, wrong logger")
logging::loginfo("This appears in logfile.txt", logger = 'main')
}
\author{
Sebastian Schweer
}
\keyword{logging}
